requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Security-Too==5.3.3
Flask-SocketIO==5.3.6
Werkzeug==2.3.6
python-dotenv==1.0.0
gevent==22.10.2
gevent-websocket==0.10.1
